API: Refactor: Move non-public files to private subdirectory
