+++ src/nvim/api/buffer.c
@@ -4,8 +4,8 @@
 #include <stdlib.h>
 
 #include "nvim/api/buffer.h"
-#include "nvim/api/helpers.h"
-#include "nvim/api/defs.h"
+#include "nvim/api/private/helpers.h"
+#include "nvim/api/private/defs.h"
 #include "nvim/vim.h"
 #include "nvim/buffer.h"
 #include "nvim/memline.h"
