+++ src/nvim/api/vim.c
@@ -4,8 +4,8 @@
 #include <string.h>
 
 #include "nvim/api/vim.h"
-#include "nvim/api/helpers.h"
-#include "nvim/api/defs.h"
+#include "nvim/api/private/helpers.h"
+#include "nvim/api/private/defs.h"
 #include "nvim/api/buffer.h"
 #include "nvim/vim.h"
 #include "nvim/buffer.h"
